+++ apps/fluent-tester/src/FluentTester/FluentTester.tsx
@@ -87,11 +87,14 @@ export const FluentTester: React.FunctionComponent<FluentTesterProps> = (props: 
 
     return (
       <View style={fluentTesterStyles.header}>
-        <Text style={[fluentTesterStyles.testHeader]} variant="heroLargeSemibold" color={theme.host.palette?.TextEmphasis}>
+        <Text
+          testID={BASE_TESTPAGE}
+          style={[fluentTesterStyles.testHeader]}
+          variant="heroLargeSemibold"
+          color={theme.host.palette?.TextEmphasis}
+        >
           ⚛ FluentUI Tests
         </Text>
-        {/* Workaround for testID prop on text component affecting text size */}
-        <Text testID={BASE_TESTPAGE}> </Text>
         <ThemePickers />
       </View>
     );