+++ packages/components/Avatar/CHANGELOG.md
@@ -1,8 +1,107 @@
 # Change Log - @fluentui-react-native/avatar
 
-This log was last generated on Tue, 28 Jun 2022 16:48:37 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Jul 2022 20:03:37 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.13
+
+Thu, 14 Jul 2022 20:03:37 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/badge to v0.1.1
+
+## 1.1.12
+
+Thu, 14 Jul 2022 18:09:50 GMT
+
+### Patches
+
+- updated to rn 0.66 (duomaster01@gmail.com)
+- Bump @fluentui-react-native/adapters to v0.9.0
+- Bump @fluentui-react-native/badge to v0.1.0
+- Bump @fluentui-react-native/framework to v0.8.0
+- Bump @fluentui-react-native/icon to v0.13.0
+- Bump @fluentui-react-native/interactive-hooks to v0.17.0
+- Bump @fluentui-react-native/theme-tokens to v0.19.0
+- Bump @fluentui-react-native/tokens to v0.16.0
+- Bump @fluentui-react-native/use-styling to v0.9.0
+
+## 1.1.11
+
+Thu, 14 Jul 2022 17:13:16 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/badge to v0.0.50
+
+## 1.1.10
+
+Tue, 12 Jul 2022 17:07:10 GMT
+
+### Patches
+
+- UI fixes (v.kozova13@gmail.com)
+- Bump @fluentui-react-native/badge to v0.0.49
+
+## 1.1.9
+
+Mon, 11 Jul 2022 15:40:06 GMT
+
+### Patches
+
+- Fixed bug with text-scaling (v.kozova13@gmail.com)
+
+## 1.1.8
+
+Fri, 08 Jul 2022 21:23:37 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/badge to v0.0.48
+- Bump @fluentui-react-native/framework to v0.7.32
+- Bump @fluentui-react-native/icon to v0.12.3
+- Bump @fluentui-react-native/interactive-hooks to v0.16.5
+- Bump @fluentui-react-native/tokens to v0.15.1
+
+## 1.1.7
+
+Fri, 08 Jul 2022 17:53:30 GMT
+
+### Patches
+
+- Fixed colorful icons in HC (vkozlova@microsoft.com)
+- Bump @fluentui-react-native/badge to v0.0.47
+
+## 1.1.6
+
+Thu, 07 Jul 2022 21:24:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/badge to v0.0.46
+- Bump @fluentui-react-native/framework to v0.7.31
+- Bump @fluentui-react-native/icon to v0.12.2
+- Bump @fluentui-react-native/interactive-hooks to v0.16.4
+- Bump @fluentui-react-native/tokens to v0.15.0
+
+## 1.1.5
+
+Thu, 07 Jul 2022 12:16:12 GMT
+
+### Patches
+
+- Fixed badge position (vkozlova@microsoft.com)
+- Bump @fluentui-react-native/badge to v0.0.45
+
+## 1.1.4
+
+Tue, 28 Jun 2022 22:01:06 GMT
+
+### Patches
+
+- Update package.json version (ruaraki@microsoft.com)
 
 ## 1.1.3