+++ packages/framework/framework/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Wed, 25 May 2022 18:43:09 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Jul 2022 18:09:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.0
+
+Thu, 14 Jul 2022 18:09:51 GMT
+
+### Minor changes
+
+- Update to React Native 0.66 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/composition to v0.8.0
+- Bump @fluentui-react-native/default-theme to v0.14.0
+- Bump @fluentui-react-native/merge-props to v0.5.0
+- Bump @fluentui-react-native/theme-types to v0.18.0
+- Bump @fluentui-react-native/use-styling to v0.9.0
+- Bump @fluentui-react-native/use-slot to v0.3.0
+- Bump @fluentui-react-native/use-slots to v0.7.0
+- Bump @fluentui-react-native/use-tokens to v0.3.0
+- Bump @fluentui-react-native/tokens to v0.16.0
+
+## 0.7.32
+
+Fri, 08 Jul 2022 21:23:37 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tokens to v0.15.1
+
+## 0.7.31
+
+Thu, 07 Jul 2022 21:24:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tokens to v0.15.0
 
 ## 0.7.30