+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,33 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Mon, 13 Jun 2022 23:58:18 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Jul 2022 18:09:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.0
+
+Thu, 14 Jul 2022 18:09:51 GMT
+
+### Minor changes
+
+- Update to React Native 0.66 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/icon to v0.13.0
+
+## 0.16.5
+
+Fri, 08 Jul 2022 21:23:37 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.12.3
+
+## 0.16.4
+
+Thu, 07 Jul 2022 21:24:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.12.2
 
 ## 0.16.3